requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
Authlib==1.6.9
cffi==2.0.0
cryptography==46.0.6
fastapi==0.135.1
pandas==2.3.3
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
requests==2.32.5
setuptools==82.0.1
sqlparse==0.5.5
starlette==0.52.1
streamlit==1.54.0
python-multipart==0.0.26
uvicorn==0.42.0
